+++ zetaclient/orchestrator/contextupdater.go
@@ -21,6 +21,7 @@ type Zetacore interface {
 	GetChainParams(ctx context.Context) ([]*observertypes.ChainParams, error)
 	GetTSS(ctx context.Context) (observertypes.TSS, error)
 	GetKeyGen(ctx context.Context) (observertypes.Keygen, error)
+	GetOperationalFlags(ctx context.Context) (observertypes.OperationalFlags, error)
 }
 
 var ErrUpgradeRequired = errors.New("upgrade required")
@@ -54,7 +55,12 @@ func UpdateAppContext(ctx context.Context, app *zctx.AppContext, zc Zetacore, lo
 
 	crosschainFlags, err := zc.GetCrosschainFlags(ctx)
 	if err != nil {
-		return errors.Wrap(err, "unable to fetch crosschain flags from zetacore")
+		return errors.Wrap(err, "unable to fetch crosschain flags")
+	}
+
+	operationalFlags, err := zc.GetOperationalFlags(ctx)
+	if err != nil {
+		return errors.Wrap(err, "unable to fetch operational flags")
 	}
 
 	freshParams := make(map[int64]*observertypes.ChainParams, len(chainParams))
@@ -86,6 +92,7 @@ func UpdateAppContext(ctx context.Context, app *zctx.AppContext, zc Zetacore, lo
 		additionalChains,
 		freshParams,
 		crosschainFlags,
+		operationalFlags,
 	)
 }